+++ regression_test.go
@@ -1,0 +1,38 @@
+package main
+
+import (
+	_ "embed"
+	"testing"
+
+	select_test "github.com/ncruces/wasm2go/testdata/regression/select_effect"
+	store_grow_test "github.com/ncruces/wasm2go/testdata/regression/store_grow"
+)
+
+func Test_regression_select_effect(t *testing.T) {
+	m := select_test.New()
+
+	if got := m.Xtest(0); got != 5 {
+		t.Errorf("test(0) = %d, want 5", got)
+	}
+	if got := m.Xcounter(); got != 1 {
+		t.Errorf("counter = %d after test(0), want 1 (operand must run unconditionally)", got)
+	}
+
+	if got := m.Xtest(1); got != 100 {
+		t.Errorf("test(1) = %d, want 100", got)
+	}
+	if got := m.Xcounter(); got != 2 {
+		t.Errorf("counter = %d after test(1), want 2", got)
+	}
+}
+
+func Test_regression_store_grow(t *testing.T) {
+	m := store_grow_test.New()
+
+	if got := m.Xtest(); got != 12345 {
+		t.Errorf("test() = %d, want 12345 (store must use memory slice after memory.grow)", got)
+	}
+	if got := m.Xsize(); got != 2 {
+		t.Errorf("size = %d, want 2", got)
+	}
+}
