+++ testdata/loops/loops.go
@@ -50,9 +50,9 @@ l1:
 			goto l0
 		}
 		v3 = v0 + v2*i32(4)
-		t1 := v4
-		t0 := int32(load32((*m.memory)[uint32(v3):]))
-		v4 = t1 + t0
+		t0 := v4
+		t1 := int32(load32((*m.memory)[uint32(v3):]))
+		v4 = t0 + t1
 		v2 = v2 + i32(1)
 		goto l1
 	}
